# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MOVE Apayao — Transit & Trip Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    post:
      operationId: createTrip
      tags: [trips]
      summary: Request a trip (passenger only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripBody"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
    get:
      operationId: listTrips
      tags: [trips]
      summary: Get trip history
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trip list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTripsResponse"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip details
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /trips/{id}/accept:
    post:
      operationId: acceptTrip
      tags: [trips]
      summary: Driver accepts a trip
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
  /trips/{id}/decline:
    post:
      operationId: declineTrip
      tags: [trips]
      summary: Driver declines a trip
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip declined and re-queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
  /trips/{id}/complete:
    post:
      operationId: completeTrip
      tags: [trips]
      summary: Driver marks trip complete and deducts fare
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
  /trips/{id}/cancel:
    post:
      operationId: cancelTrip
      tags: [trips]
      summary: Cancel a trip
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Trip cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
  /trips/{id}/rate:
    post:
      operationId: rateTrip
      tags: [trips]
      summary: Rate a completed trip
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateTripBody"
      responses:
        "201":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripRatingResponse"

  # ── Loop Routes ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Trip:
      type: object
      properties:
        id:
          type: integer
        passengerId:
          type: string
          format: uuid
        driverId:
          type: string
          format: uuid
          nullable: true
        pickupLat:
          type: number
        pickupLon:
          type: number
        pickupAddress:
          type: string
          nullable: true
        dropoffLat:
          type: number
        dropoffLon:
          type: number
        dropoffAddress:
          type: string
          nullable: true
        status:
          type: string
          enum: [requested, matched, in_progress, completed, cancelled]
        fareAmount:
          type: number
          nullable: true
        distanceKm:
          type: number
          nullable: true
        cancelReason:
          type: string
          nullable: true
        matchedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, passengerId, pickupLat, pickupLon, dropoffLat, dropoffLon, status, createdAt]
    TripResponse:
      type: object
      properties:
        trip:
          $ref: "#/components/schemas/Trip"
      required: [trip]
    ListTripsResponse:
      type: object
      properties:
        trips:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
        limit:
          type: integer
        offset:
          type: integer
      required: [trips, limit, offset]
    CreateTripBody:
      type: object
      properties:
        pickupLat:
          type: number
        pickupLon:
          type: number
        pickupAddress:
          type: string
        dropoffLat:
          type: number
        dropoffLon:
          type: number
        dropoffAddress:
          type: string
      required: [pickupLat, pickupLon, dropoffLat, dropoffLon]
    TripRating:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        raterId:
          type: string
          format: uuid
        rateeId:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, tripId, raterId, rateeId, rating, createdAt]
    TripRatingResponse:
      type: object
      properties:
        rating:
          $ref: "#/components/schemas/TripRating"
      required: [rating]
    RateTripBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [rating]

    # ── Loop Routes ──────────────────────────────────────────────────────────
...